// src/constants/interfaces.ts
export interface KV<U extends Record<string, any>> {
  key: string;
  values: U[];
  label: keyof U;
}

export const graphTypes = ["pie", "line", "bar", "area"] as const;

export type GraphType = (typeof graphTypes)[number];

export type ActionModalDef = {
  icon: string;
  label: string;
  name: string;
  modal: React.ReactNode;
};
// ...
export type PaginatedResponse<T> = {
  count: number;
  next: string | null;
  previous: string | null;
  results: T[];
  currentPage: number;
  totalPages: number;
  ids: number[];
};

export type PaginatedDetails = Omit<PaginatedResponse<unknown>, "results">;

export type Graph = "line" | "pie";

export type StateSetter<T> = React.Dispatch<React.SetStateAction<T>>;